fcntl()'s third argument is an int (except when it's an intptr_t) per the manpage.

+++ lib/libc/sys/fcntl.c
@@ -70,7 +70,7 @@ _fcntl(int fd, int cmd, ...)
 		break;
 
 	default:
-		arg = va_arg(args, long);
+		arg = va_arg(args, int);
 		break;
 	}
 	va_end(args);
